+++ lib/regax_validator.dart
@@ -100,3 +100,4 @@ class RegexValidator {
     return regex.hasMatch(email.trim());
   }
 }
+////////ddddd